Base template of creating views that execute and export csv files.
Creating a python script that runs sql statements that create views of data from two tables. Gets these views as a dataframe and then exports this as a flat file. Tests still need to be done for execute and export functions. Will need to be heavily adapted for Snowflake and test data once accessed is received.