Updated sql statements in dbt for nation and region hubs.
There are two feeds to the nation and region (from customer and supplier). An sql statement has been added that creates a intermediate view between the stage and the hub. These hubs then load from these views which are then dropped after their use. If the demo is being run with the tpch data these hubs will need to be hard-coded into the dbt folder as the hub template is different. Next commit looks to integrate parallel processing to speed up the demo.