Addition of additional data files and improvement to satellite statement
Template Generator now supports the option of writing hubs, links, and sat metadata is .data files to avoid config becoming too big a file. If no additional files provided it will run as normal. Change to logic of satellite statement in dbt, seems to work as intended but will add tests soon and then add these logic changes to the hubs and links.