Adjustments for two source feeds made.
The sample data has now been split to represent the 2 source feeds. Each source now has a staging table and the hubs and links are loaded to from the correct staging areas. Next is to create a a usage link and satellite, then create bridges and build the star schema.