Added md5 binary macro templates and modified stage template.
Changes to the stage template, it now uses two different macros for calculating the hashes. It also generates these macros as well, making for cleaner code. Removal of some redundant files. Update the generator to take the ~ character in the config and replace it with a comma for the data types.