Orchestrator now logs properly.
The problem with the orchestrator not logging properly now fixed, now logs to a separate directory to dbt. Unit test added to check for proper logging. Disabled the dbt project yaml file generation as it is to specific to the TPCH data. Once generalisation has been done on this feature where a specific yaml file can be constructed from metadata supplied by the config.