Beginning of restructuring of tests.
Beginning of restructuring of tests to enables testing of a deprecated version of dbtvault and a current version of dbt vault. Currently, the deprecated version of load_hubs passes. Next steps: get other feature files working in the deprecated version first, then work on the getting the current version tests working.