Improved Eff Sat to exclude NULL SFKs and closed previous entries
The eff sat now does not add new entries if the secondary foreign key is NULL. If a secondary foreign key is NULL then it closes the most recent eff sat entry with the common driving key. Eff sat feature tests currently all pass. Link macro needs to adjust for this and maybe another scenario written for the link.