Refactoring of eff sat and link macro.
Eff sat and link macros refactored to account for the case of one of the FK's being NULL. If one of the FK's is NULL, the link no longer inserts a new record. This currently only works for the link for incremental runs, this may change in the future to be included in initial runs. A new macro was created called hash_check which gets the required hashing algorithm to check for NULLS in the hashed FKs.