Further refactoring and improving eff sat and link macro.
The eff sat macro now does not insert a new row if the driving foreign key is null, however it does not close any existing ones if the dfk is NULL, it only closes the latest eff sat entries if the sfk is NULL. The Link macro now Does not create new links if either of the foreign keys are NULL, including for base loads for single and union, as well as union incremental loads. Several more tests have been added to the link feature files to test this. A test has been added in the eff sat feature file to test the dfk being NULL. A refactor of the link macro is required as currently it is a little messy.