Eff sat macro refactor.
The eff sat macro code has been refactored and cleaned up. This has been done by creating a new macro called multikey. This macro takes in the driving keys or secondary keys, any aliases in a list, and the type of for loop need. Based off the input parameters, it writes the required sql. May need a few adjustments to clean up compiled code, however this is not essential.