Beginning of eff sat documentation.
A eff sat section has been added in the macro documentation. The first draft of the eff sat documentation has been done for the walk through guide. An example table of data is still needed for this section. A section needs to be added to the worked example and maybe sections added to the staging in the walk through example.