Refactored Eff sat macro
Eff sat macro did not work in snowflake demo when trying to create a eff sat that didn't exist. This was due to the WITH clause being outside of the incremental section (dbt still tries to find dependencies in WITH clauses even if they are not used). An additional change has to be made to the model as well for this macro to work properly.