Added lowest source feature into hub_macro/new_union macro
For the new hub macro, when a union occurs if there is a union record that agrees on the earliest date, only the lowest source will be inserted into the hub. All hub tests current pass. A new test is needed to test the different variations of NULL values that could occur.