Refactor of new link sql
The new link sql has been refactored into a similar format as the hubs. The multiperiod_link macro has been removed as new link sql is now in the link macro. All feature file tests for links pass for the single day and multiperiod loads. Next steps to add a union test for the multipart keys for the links and the hubs.