WIP: Exclude flag in hashdiff during staging
Macro for loading the data for use in the hashdiff that is not in the exclude flag. Minor changes to other files. Comit before refactoring the stage macro and helper macros to use the exclude flag in hashdiff. 
WIP: Exclude flag in hashdiff during staging

renamed the calculate_hash_exclusions to process_exclude and made adjustments. Currently when the exclude flag is true an empty list is returned rather than the columns we do want to hash. Models for the unit tests where also added.
WIP: Exclude flag in hashdiff during staging 

refactored process_exclude macro and started taking logic out and putting it into its own macro in the same file. Added a unit test for exclude flag for hashdiff.