[FEATURES] Staging Quality of Life (#1)
* Fixed seed config


Moved comments in test output


Minor

* Added gitignore

* WIP: Start of stage helper macros refactor

* Update to dbt v0.18.1

* Refactor of stage helper macros

* Removed repeated logic in the helper stage macros

* WIP: Adding unit tests for exclude flag in hashdif during staging 

Only added for stage tests but none for tests that specifically test hashed columns exclusively

* WIP: Exclude flag in hashdiff during staging

Macro for loading the data for use in the hashdiff that is not in the exclude flag. Minor changes to other files. Comit before refactoring the stage macro and helper macros to use the exclude flag in hashdiff.

* WIP: Exclude flag in hashdiff during staging

renamed the calculate_hash_exclusions to process_exclude and made adjustments. Currently when the exclude flag is true an empty list is returned rather than the columns we do want to hash. Models for the unit tests where also added.

* WIP: Exclude flag in hashdiff during staging 

refactored process_exclude macro and started taking logic out and putting it into its own macro in the same file. Added a unit test for exclude flag for hashdiff.

* WIP: Exclude flag in hashdiff during staging

Macro for loading the data for use in the hashdiff that is not in the exclude flag. Minor changes to other files. Comit before refactoring the stage macro and helper macros to use the exclude flag in hashdiff. 
WIP: Exclude flag in hashdiff during staging

renamed the calculate_hash_exclusions to process_exclude and made adjustments. Currently when the exclude flag is true an empty list is returned rather than the columns we do want to hash. Models for the unit tests where also added.
WIP: Exclude flag in hashdiff during staging 

refactored process_exclude macro and started taking logic out and putting it into its own macro in the same file. Added a unit test for exclude flag for hashdiff.

* Added exclude_hashdiff prefix capability for context tables

* changed exclude_flag when hashing to exclude_columns.

Process exclude macro was also refactored and added feature test scenarios.

* Added setup task to improve setup time

* Added change_target task

* All unit and feature tests passing for staging 

after implementation of an exclude flag for calculating hashdiff's

* Minor formatting to Stage macros

* Fixed run-dbt task and renamed dev package folder

* Added template file override for contributors

* Update circleci config

* Updated contrib and READMEs

* Formatting fixes

* Fixed failing derived columns test

* Cleanup gitignores

* Changed test project.yml back to non-generated

* Added after_all to environment for feature tests


Formatting fix


Fix test

* Stage macro cleanup

* Changed dummy variable list updates to 'do' syntax


Change process_source_and_derived macro name

* Updated contrib readme

* Minor formatting fix

* Updated branch config
Fix

* Formatting fix for dbt_project.yml

Co-authored-by: Alex Higgs <alex.higgs@data-vault.co.uk>