WIP: Bridge macro bug fixes - see comments
eff sat start date <= as of date
eff sat end date > as of date
The eff sat start date, when using auto end dating, equals the load date time, however, when not using auto end dating we must use the start date and not the load date time as the prevailing business rules will be determining the start date on some basis. So always use the eff sat start date to determine relationship currency for each bridge as of date.