WIP *Where clause removed temp*
In the new_rows_as_of CTE it appears the WHERE clause it not functioning correctly, probably due to issues with the date types.  This was commented out, leading to the bridge expected vs actual tests to pass.   The date type of as_of and last_safe_load_datetime should be investigated.