">" comparison used in satellite_update CTE to compare load dates
In order for each rank/period materialisation to be loaded correctly it is required that this comparison operator is ">" oppose to ">=".  However, this means for test scenarios that have subsequent loads which contain new records with the same load datetime as previous loads are not inserted.  See comment in "one_cdk_base_sats_rank_mat_bigquery.feature" for more info.