Bug fix for oos matching record selection logic
When checking the existing records before and after the oos insertion point whilst the load dates of the preceding and subsequent records were begin correctly calculated the preceding record hashdiff was being calculated as the hashdiff of the record preceding the record preceding the insertion point. The preceding record hashdiff calculation has been corrected to use the same record as for the preceding record load date calculation. The subsequent record hashdiff calculation was already correct.