sat: Correction for oos matching record selection logic (#72)
* created test SF-SAT-OOS-11

This test currently fails due to the way the previous hash-diff is calculated using a lag in the oos logic

* Bug fix for oos matching record selection logic

When checking the existing records before and after the oos insertion point whilst the load dates of the preceding and subsequent records were begin correctly calculated the preceding record hashdiff was being calculated as the hashdiff of the record preceding the record preceding the insertion point. The preceding record hashdiff calculation has been corrected to use the same record as for the preceding record load date calculation. The subsequent record hashdiff calculation was already correct.

* Add sats + oos feature files to pipe line

* Pipeline config cleanup

Co-authored-by: Flynn Sheridan <Flynn.sheridan@data-vault.com>
Co-authored-by: Alex Higgs <alex.higgs@data-vault.com>